test(user): add user password change service and improve use cases
