+++ src/main/java/com/ironcore/application/user/usecase/InitialChangePasswordResult.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 public record InitialChangePasswordResult(
         String accessToken,
         String tokenType,
-        LocalDateTime expiredAt
+        LocalDateTime expiresAt
 ) {
 }
